Removed localization of keywords, added header.
Localization of values only provides a performance benefit when accessing functions in libraries that are represented by a table. This is because the following code...
```lua
math.max(2, 3)
```
...is syntactic sugar for looking up "max" inside a table stored at "math" in the global scope. In other words, math.max represents two table lookups - globalTable["math"] and math["max"] - whereas localizing like so...
```lua
local max = math.max
max(2, 3)
```
...does the first lookup once and localizes it to the variable max, skipping any future lookups.

However, localization does not provide any benefit for lua keywords / key functions because they are built into the language. In normal lua, you might get a tiny performance benefit, but Roblox's Luau optimizations have removed the need for localizing stuff like ipairs. If anything, it makes the code more difficult to read for beginners - the target audience for this module.